5.1.1
Corrected bug that caused by different hall effect sensor positions on
different boards.  -- hall effect handling can now be completely
disabled.  This is just a patch until I decide how to re-configure the
hall effect sensor on a board-to-board basis.  Also, added punctuation
as an opportunity to save to SD.