5.2.9
Slight change to how Bluetooth mode is configured.  Now it forces a
configuration when Bluetooth is tested with S2 and S3, and forces a
configuration first time B is held.  Holding B does not cause
re-configuration of module, but does cause module to clear its pairing
memory and search for a new partner to pair with.